backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
psycopg==3.1.19
psycopg-binary==3.1.19
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.1
gunicorn==20.1.0